extracting InMemoryEventStoreFactoryTest functionality into AbstractInMemoryEventStoreFactoryTest
